+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Unpublished
 ### Added
 - add setting for include member birthdays into public ics export
+- add button-dropdown menu at member birthday lists
 ### Changed
 - use standard notification message for click on calendar link on dashboard
 - store member filter selection in database instead of browser